+++ config/routes.rb
@@ -21,6 +21,7 @@ Dorsale::Engine.routes.draw do
       member do
         patch :complete
         patch :snooze
+        patch :copy
       end
     end