+++ spec/controllers/dorsale/flyboy/tasks_controller_spec.rb
@@ -234,6 +234,11 @@ describe Dorsale::Flyboy::TasksController, type: :controller do
       patch :snooze, params: {:id => task.to_param}
       expect(response).to redirect_to(task)
     end
+
+    it "is expected to create a task comment" do
+      task = create(:flyboy_task, term: 3.days.ago)
+      expect{patch :snooze, params: {:id => task.to_param}}.to change{Dorsale::Flyboy::TaskComment.count}.by(1)
+    end
   end
 
   describe "summary" do