+++ spec/dummy/db/migrate/20170522142733_enable_unaccent.agilibox.rb
@@ -1,0 +1,6 @@
+# This migration comes from agilibox (originally 20170502143330)
+class EnableUnaccent < ActiveRecord::Migration[5.0]
+  def change
+    enable_extension "unaccent"
+  end
+end